Move the fetchers into their own file
